bin: Add a tool to generate a template bootrr board description
bootrr requires a description of each board it runs on to provide
detection of the actual devices and system support that is
present.  Currently these descriptions must be written by hand
but we can ease the process of creating them by providing a tool
which examines the running system and outputs a set of bootrr
assertions which would pass on the system.

The script will require post processing by users, the generate
rule names won't be good, hotplugged devices will be included and
any spaces in device or driver names will cause confusion, but it
can provide a usefuls starting point.

Signed-off-by: Mark Brown <broonie@kernel.org>